requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.68.2
firecrawl-py==1.12.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.108.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
multidict==6.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==1.61.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.14.2
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.15.0
websockets==15.0
yarl==1.18.3
